refactoring files for next naive unify.
1. separate unify from reso.jl
2. separate type definition from reso.jl to types.jl
3. separate apply from unify.jl to subst.jl
4. separete tests for primitives, unify, unifybase, subst
5. test test_cheaplogic, test_kplogic
6. comment out include in testxxx.jl for what files loaded by control only load_xxx.jl

	modified:   load_cheaplogic.jl
	modified:   load_kplogic.jl
	modified:   reso.jl
	new file:   subst.jl
	modified:   test_cheaplogic.jl
	modified:   testdvcreso.jl
	modified:   testmerge.jl
	modified:   testnewcore.jl
	modified:   testparser.jl
	new file:   testprimitives.jl
	modified:   testreso.jl
	new file:   testsubst.jl
	new file:   testunify.jl
	new file:   testunifybase.jl
	new file:   types.jl
	new file:   unify.jl
	new file:   unifybase.jl
