+++ Prover/load_cheaplogic.jl
@@ -1,6 +1,11 @@
 include("config.jl")
-include("misc.jl")
 include("utils.jl")
+include("misc.jl")
+include("types.jl")
+include("primitives.jl")
+include("subst.jl")
+include("unifybase.jl")
+include("unify.jl")
 include("reso.jl")
 include("newcore.jl")
 include("coreprint.jl")
@@ -8,5 +13,6 @@ include("dvcreso.jl")
 include("analyzer.jl")
 include("repl.jl")
 include("merge.jl")
+include("parser.jl")
 
 println("... cheaplogic is loaded")
