+++ Prover/load_kplogic.jl
@@ -3,6 +3,9 @@ include("load_cheaplogic.jl")
 include("config.jl")
 include("misc.jl")
 include("utils.jl")
+include("types.jl")
+include("primitives.jl")
+
 #include("reso.jl")
 
 include("kptype.jl")
