+++ Prover/reso.jl
@@ -2,6 +2,8 @@
 
 include("misc.jl")
 include("primitives.jl")
+include("types.jl")
+include("unify.jl")
 
 # input form
 # "[x,y].[+P(x,f(x),-Q(x)]"
@@ -22,429 +24,6 @@ EmptyClause = []
 CForm = Tuple{Array,Array}
 EmptyCForm = ([],EmptyClause)
 
-## Exceptions
-"""
-Exception for ununifiable terms in subst
-"""
-struct ICMPIn
- left
- right
- subst
- op
-end
-
-"""
-Exception for ununifiable terms
-"""
-struct ICMP
- left
- right
- op
-end
-
-struct Loop
- left
- right
- op
-end
-
-"""
-apply substitution to a term, substitution
-"""
-function apply(vars::Vlist, sym::Number, subst::Tlist)
-#@show :applynt, sym, subst
- return sym
-end
-
-function apply(vars::Vlist, sym::Symbol, subst::Tlist)
-#@show :applyst, sym, subst
- for i in 1:length(vars)
-  if sym == vars[i]; return subst[i] end
- end
- return sym
-end
-
-"""
-apply: ΣxΣ → Σ
-"""
-function apply(vars::Vlist, subst1::Tlist, subst2::Tlist)
-#@show :applytt, subst1, subst2
- nterm = []
- for arg in subst1
-  narg = apply(vars, arg, subst2)
-  if isempty(nterm)
-   nterm = [narg]
-  elseif ndims(subst1) == 1
-     nterm = vcat(nterm, [narg])
-  else
-     nterm = hcat(nterm, [narg])
-  end
- end
- nterm
-end
-
-function apply(vars::Vlist, term::Expr, subst::Tlist)
-#@show :applyet, term, subst
- nterm = deepcopy(term)
- for i in 1:length(term.args)
-  arg = term.args[i]
-  narg = apply(vars, arg, subst)
-  nterm.args[i] = narg
- end
- nterm
-end
-
-## unification
-"""
-putasubst puts a pair of v,t to a substituion
-"""
-function putasubst(vars::Vlist, vt::Symbol, tm::Term, subst::Tlist)
- rs = map(x->if x==vt;tm else x end, subst)
-end
-
-function unify0(vars::Vlist, t1::Number, t2::Number)
- if t1==t2; return () end
- throw(ICMP(t1,t2,:unify0nn))
-end
-
-function unify0(vars::Vlist, t1::Number, t2::Symbol)
- if isvar(t1,vars); return (t1,t2) end
-# if isvar(t2,vars); return (t2,t1) end
- throw(ICMP(t1,t2,:unify0ns))
-end
-
-function unify0(vars::Vlist, t1::Number, t2::Expr)
- throw(ICMP(t1,t2,:unify0ne))
-end
-
-function unify0(vars::Vlist, t1::Expr, t2::Number)
- throw(ICMP(t1,t2,:unify0en))
-end
-
-function unify0(vars::Vlist, t1::Symbol, t2::Number)
-# if isvar(t1,vars); return (t1,t2) end
- if isvar(t2,vars); return (t2,t1) end
- throw(ICMP(t1,t2,:unify0sn))
-end
-
-function unify0(vars::Vlist, t1::Symbol, t2::Symbol)
- if t1==t2; return () end
- if isvar(t1,vars) && isvar(t2,vars)
-   if isinvar(t1); return (t2,t1) end
-   if isinvar(t2); return (t1,t2) end
- end
- if isvar(t1,vars); return (t1,t2) end
- if isvar(t2,vars); return (t2,t1) end
- if t1!=t2; throw(ICMP(t1,t2,:unify0ss)) end
-end
-
-function loopcheck(vt::Symbol, et::Number; istop=false)::Bool
-  return false
-end
-
-function loopcheck(vt::Symbol, et::Symbol; istop=false)::Bool
-  return false
-end
-
-function loopcheck(vt::Symbol, et::Expr; istop=true)::Bool
-  for arg in et.args
-   if typeof(arg) == Symbol
-    if vt == arg; throw(Loop(vt,et,:loopse)) end
-   else
-    if loopcheck(vt, arg, istop=false);return throw(Loop(vt,et,:loopse)) end
-   end
-  end
-  return false
-end
-
-function unify0(vars::Vlist, t1::Symbol, t2::Expr)
- isvar(t1,vars) && !loopcheck(t1,t2) && return (t1,t2) 
- if t1!=t2; throw(ICMP(t1,t2,:unify0se)) end
-end
-
-function unify0(vars::Vlist, t1::Expr, t2::Symbol)
- isvar(t2,vars)&&!loopcheck(t2,t1)&&return(t2,t1)
- if t1!=t2; throw(ICMP(t1,t2,:unify0es)) end
-end
-
-function unify0(vars::Vlist, t1::Expr, t2::Expr)
- if t1==t2; return () end
- for i in 1:length(t1.args)
-  if t1.args[i]==t2.args[i]; continue end
-  σi=unify0(vars,t1.args[i],t2.args[i])
-  if σi!=(); return σi end 
- end
- return ()
-end
-
-### unify do unification
-# t1,t2 are literal(without sign, without or/and.
-
-function maketlist(vars::Vlist, t1::Symbol, t2::Symbol)::Tlist
-  map(x->if x==t1;t2 else t1 end, vars)
-end
-
-#function unify1(vars::Vlist, t1::Symbol, t2::Symbol, subst::Tlist)
-##  if t1!=t2; throw(ICMP(t1,t2,:unify1)) end
-#  σb=vars
-#  while(true)
-#   σa0=unify0(vars,t1,t2)
-#   if σa0==(); break end
-#   σa = maketlist(vars, σa0[1], σa0[2]) 
-##@show σa
-#   σi::Tlist = merge(vars, σb, σa)
-##@show σi
-#   t1 = apply(t1,σi,σb)
-#   t2 = apply(t2,σi,σb)
-#   σb=σa
-#  end
-# σb
-#end
-
-function unify1(vars::Vlist, t1::Number, t2::Number, subst::Tlist)
-#@show :unify1vnn,t1,t2,subst
- if t1==t2; return subst end
- throw(ICMP(t1,t2,:unify1nn))
-end
-
-function unify1(vars::Vlist, t1::Expr, t2::Number, subst::Tlist)
-#@show :unify1ven,t1,t2,subst
- throw(ICMP(t1,t2,:unify1en))
-end
-function unify1(vars::Vlist, t1::Number, t2::Expr, subst::Tlist)
-#@show :unify1vne,t1,t2,subst
- throw(ICMP(t1,t2,:unify1ne))
-end
-
-"""
-replace the ix'th element of tlist with t
-without typechecking
-"""
-function putarray(tlist::Array, ix::Number, t::Any)
-  map(i->if i==ix; t else tlist[i] end, 1:length(tlist))
-end
-
-function vindex(vars::Vlist, v::Symbol)
- ix = findfirst(x->x==v,vars)
- if ix == nothing; return 0 end
- return ix
-end
-
-"""
-replace v with t in σ when v is var. and fixed point of σ.
-if v<-s in σ, and s is var, replace v with t.
-otherwise exception
-"""
-function putsubst(vars, v::Symbol, t::Any, subst::Tlist)
- if v == t; return subst end
-
- ix = vindex(vars, v)
- if ix == 0; println("unknown var $v in putsubst");return subst end
-
- ot = subst[ix]
- if ot == t; return subst end
- if !isvar(ot, vars)
-  if ot == t; return subst
-  elseif isvar(t, vars)
-    v = t
-    t = ot
-    ix = vindex(vars, v)
-    if ix == 0; throw(ICMP(ot,t,:putsubst)) end
-  else
-   throw(ICMP(ot,t,:putsubst)) 
-  end
-
-  subst=putarray(subst, ix, t)
-  return subst
- else
-  subst=putarray(subst, ix, t)
-
-  rsubst=[]
-  while subst != rsubst
-   ix = vindex(vars, ot)
-   if ix == 0; println("unknown var $v in putsubst")
-     throw(ICMP(ot,t,:unify1ss3))
-   end
-#@show ix
-   rsubst=putsubst(vars, ot, t, subst) # backward subst
-   subst=rsubst
-#@show subst,rsubst
-   ot = subst[ix]
-  end
-  fsubst = fp_subst(vars, rsubst) # forward subst
-  return fsubst
- end
-#@show "$ot is not ovar"
- if t == ot; return fp_subst(vars, subst) end
- throw(ICMP(v,t,:unify1ne))
-end
-
-function unify1(vars::Vlist, t1::Symbol, t2::Number, subst::Tlist)
-#@show :unify1vsn,t1,t2,subst
- putsubst(vars, t1, t2, subst)
-end
-
-function unify1(vars::Vlist, t1::Number, t2::Symbol, subst::Tlist)
-#@show :unify1vns,t1,t2,subst
- putsubst(vars, t2, t1, subst)
-end
-
-function unify1(vars::Vlist, t1::Symbol, t2::Symbol, subst::Tlist)
-#@show :unify1vss,t1,t2,subst
- if t1==t2;return subst end
-
- if isvar(t1,vars)
-   return putsubst(vars, t1, t2, subst)
- end
- if isvar(t2,vars)
-   return putsubst(vars, t2, t1, subst)
- end
-
- throw(ICMP(t1,t2,:unify1ss3))
-# if isvar(t1,vars)
-#  ix = findfirst(x->x==t1,vars)
-#  if t1 == subst[ix] # t1 is var and havn't substituted yet. 
-#    subst[ix] = t2
-#  elseif isvar(subst[ix],vars)
-#@show :unify1vss1, vars, t1, t2, subst[ix], vars
-##    if t2!=subst[ix]; throw(ICMPIn(t1,t2,subst,:unify1ss1))end 
-#    subst[ix] = t2
-#  end
-#  return subst
-# end
-#
-# if isvar(t2,vars)
-#  ix = findfirst(x->x==t2,vars)
-#  if isvar(subst[ix], vars)
-#@show :unify1vss2, vars, t1, t2, subst[ix], vars
-#    subst[ix] = t1
-#  end
-#  return subst
-# end
-# throw(ICMP(t1,t2,:unify1ss3))
-end
-
-function unify1(vars::Vlist, t1::Symbol, t2::Expr, subst::Tlist)
-#@show :unify1vse,t1,t2,subst
- if isvar(t1,vars);return putsubst(vars, t1, t2, subst) end
-# if isvar(t1,vars);return t2 end
- throw(ICMP(t1,t2,:unify1se))
-end
-
-function unify1(vars::Vlist, t1::Expr, t2::Symbol, subst::Tlist)
-#@show :unify1ves,t1,t2,subst
- if isvar(t2,vars);return putsubst(vars, t2, t1, subst) end
-# if isvar(t2,vars);return t2 end
- throw(ICMP(t1,t2,:unify1))
-end
-
-"""
-unify1 has a var subst for internal use
-"""
-function unify1(vars::Vlist, t1::Expr, t2::Expr, subst::Tlist)
-#@show :unify1vee,t1,t2,subst
- if t1==t2; 
-  return subst
- end
-
- if t1.args[1] != t2.args[1];
-  throw(ICMP(t1,t2,:unify1ee_1))
- end 
-
- if length(t1.args) != length(t2.args); return vars end
- while true  
-  pp = unify0(vars, t1, t2)
-  if isempty(pp); break end
-  subst = putasubst(vars, pp[1], pp[2], subst)
-  t1 = apply(vars,t1,subst)
-  t2 = apply(vars,t2,subst)
- end
- subst
-end
-
-"""
-unify is a main interface for unification
-"""
-
-function loopcheck_sigma(vars::Vlist, subst::Tlist)
-  for ix in 1:length(vars)
-    if typeof(subst[ix]) != Symbol
-      loopcheck(vars[ix], subst[ix])
-    end
-  end
-  return false
-end
-
-function fp_subst(vars::Vlist, subst::Tlist)
-#@show vars, subst
-## make a fixed point of sigma(ss)
- sb = subst
- sa = apply(vars, subst, subst)
- while sb != sa
-   loopcheck_sigma(vars, sa)
-   sb, sa = sa, apply(vars, sa, sa)
- end
- return sa
-end
-
-function fp_subst2(vars::Vlist, subst::Tlist)
-## make a fixed point of sigma(ss)
- sb = subst
- sa = apply(vars, subst, subst)
- return sa
-end
-
-function unify(vars::Vlist, t1::Number, t2::Number)
-#@show :unifynn,t1,t2
-  if t1 != t2; throw(ICMP(t1,t2,:unify1nn))end
-  vars
-end
-
-function unify(vars::Vlist, t1::Number, t2::Symbol)
-#@show :unifyns,t1,t2
-  if t1 == t2; return vars end
-  unify1(vars, t1, t2, vars)
-end
-
-function unify(vars::Vlist, t1::Symbol, t2::Number)
-#@show :unifysn,t1,t2
-  if t1 == t2; return vars end
-  unify1(vars, t1, t2, vars)
-end
-
-function unify(vars::Vlist, t1::Symbol, t2::Symbol)
-#@show :unifyss,t1,t2
-  if t1 == t2; return vars end
-  unify1(vars, t1, t2, deepcopy(vars))
-end
-
-function unify(vars::Vlist, t1::Number, t2::Expr)
-#@show :unifyne,t1,t2
-  unify1(vars, t1, t2, vars)
-end
-
-function unify(vars::Vlist, t1::Symbol, t2::Expr)
-#@show :unifyse,t1,t2
-  unify1(vars, t1, t2, vars)
-end
-
-function unify(vars::Vlist, t1::Expr, t2::Number)
-#@show :unifyen,t1,t2
-  unify1(vars, t1, t2, vars)
-end
-
-function unify(vars::Vlist, t1::Expr, t2::Symbol)
-#@show :unifyes,t1,t2
-  unify1(vars, t1, t2, vars)
-end
-
-function unify(vars::Vlist, t1::Expr, t2::Expr)
-#@show :unifyee,t1,t2
- ss = unify1(vars, t1, t2, vars)
-#@show vars, ss
- fs = fp_subst(vars, ss)
- return fs 
-end
 
 ## resolution of 2 clauses
 function genvars(vars::Vlist)::Vlist
@@ -608,17 +187,6 @@ function containvar(vars::Vlist, cls::Clause)
   return union(cvars,cvars)
 end
 
-function reducesub(vars::Vlist, fvars::Vlist, sub::Tlist)
- fsub  = []
- if isempty(sub); return [] end
- for i in 1:length(vars)
-  if vars[i] in fvars
-    push!(fsub, sub[i])
-  end
- end
- return fsub
-end
-
 function fitvars(vars::Vlist, cls::Clause, sub::Tlist)
  fvars = containvar(vars, cls)
  fsub = reducesub(vars, fvars, sub)
@@ -650,5 +218,16 @@ function equalclause(clause1::CForm, clause2::CForm)
  scls2 = apply(var2, cls2, var1)
  if cls1 == scls2; return true end
  return false
+end
+
+function reducesub(vars::Vlist, fvars::Vlist, sub::Tlist)
+ fsub  = []
+ if isempty(sub); return [] end
+ for i in 1:length(vars)
+  if vars[i] in fvars
+    push!(fsub, sub[i])
+  end
+ end
+ return fsub
 end
