+++ Prover/test_cheaplogic.jl
@@ -1,7 +1,11 @@
 include("load_cheaplogic.jl")
 
+
+include("testprimitives.jl")
 include("testreso.jl")
 include("testutils.jl")
+include("testunifybase.jl")
+include("testunify.jl")
 include("testmerge.jl")
 include("testnewcore.jl")
 include("testdvcreso.jl")
