+++ Prover/testprimitives.jl
@@ -1,0 +1,28 @@
+# test for primitives.jl
+using Test
+#include("primitives.jl")
+
+@testset "vars and const" begin
+ @test isvar(:x, [:y,:x]) == true
+ @test isvar(:y, [:y,:x]) == true
+ @test isvar(:z, [:y,:x]) == false
+ @test isvar(12, [:y,:x]) == false
+ @test isvar(23.1, [:y,:x]) == false
+ @test isvar(0, [:y,:x]) == false
+ @test isvar([:a :b], [:y,:x]) == false
+ @test isvar([], [:y,:x]) == false
+
+ @test isevar(:Abc, [:Abc, :x]) == true
+ @test isevar(:Abc, [:abc, :x]) == false
+ @test isevar(:dbc, [:abc, :x]) == false
+
+ @test isconst(:x, [:y,:x]) == false
+ @test isconst(:a, [:y,:x]) == true
+ @test isconst(12, [:y,:x]) == true
+ @test isconst(23.1, [:y,:x]) == true
+ @test isconst(0, [:y,:x]) == true
+ @test isvar([:a :b], [:y,:x]) == false
+ @test isvar([], [:y,:x]) == false
+
+end
+
